+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ## Release Notes
 
 - Add theoretical and confirmed memory leak hotfixes. [issue](https://github.com/archfz/cypress-terminal-report/issues/296)
+- Update cypress to 15.4.0 to confirm support.
 
 #### 7.2.1